+++ Electron/vue3/packages/renderer/src/TUIRoom/components/ManageMember/MemberItem/useMemberItemHooks.ts
@@ -1,65 +1,24 @@
-import { ref, nextTick } from 'vue';
+import { computed, ref } from 'vue';
 import { useRoomStore } from '../../../stores/room';
 import { storeToRefs } from 'pinia';
 
-
-export default function useMemberItem() {
+const showUserId = ref(''); // 需要展示操作面板的用户id
+export default function useMemberItem(userId: string) {
   const roomStore = useRoomStore();
   const { isMaster } = storeToRefs(roomStore);
-
-  const showMemberControl = ref(false);
-
-  function handleMouseEnter() {
-    if (isMaster.value) {
-      showMemberControl.value = true;
-    }
+  const isMemberControlAccessible = computed(() => (userId === showUserId.value) && isMaster.value); // 只有房主才打开操作面板
+  function openMemberControl() {
+    showUserId.value = userId;
   }
 
-  function handleMouseLeave() {
-    showMemberControl.value = false;
+  function closeMemberControl() {
+    showUserId.value = '';
   }
-
-  const memberItemEl = ref();
-  function setMemberItemRef(el: any) {
-    memberItemEl.value = el;
-  }
-
-  const memberInfoEl = ref();
-  const MemberControlEl = ref();
-  function setMemberInfoRef(el: any) {
-    memberInfoEl.value = el;
-  }
-  function setMemberControlRef(el: any) {
-    MemberControlEl.value = el;
-  }
-
-  function handleMemberItemClick() {
-    if (isMaster.value && !showMemberControl.value) {
-      showMemberControl.value = true;
-    }
-  }
-  async function handleCloseControl() {
-    await nextTick();
-    showMemberControl.value = false;
-  }
-
-  function handleDocumentClick(event: MouseEvent) {
-    if (showMemberControl.value && memberItemEl.value && !memberItemEl.value.contains(event.target)) {
-      showMemberControl.value = false;
-    }
-  }
-
 
   return {
-    showMemberControl,
-    handleMouseEnter,
-    handleMouseLeave,
-    setMemberItemRef,
-    setMemberInfoRef,
-    setMemberControlRef,
-    handleMemberItemClick,
-    handleDocumentClick,
-    handleCloseControl,
+    isMemberControlAccessible,
+    openMemberControl,
+    closeMemberControl,
   };
 }