+++ Electron/vue3/packages/renderer/vite.config.ts
@@ -1,12 +1,8 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import resolve from 'vite-plugin-resolve'
-import AutoImport from 'unplugin-auto-import/vite'
 import electron from 'vite-plugin-electron/renderer'
-import { ElementPlusResolver } from 'unplugin-vue-components/resolvers'
-import Components from 'unplugin-vue-components/vite'
 import pkg from '../../package.json'
-import { createSvg } from './src/TUIRoom/assets/icons/index.js'
 
 const path = require('path')
 
@@ -82,15 +78,6 @@ export default defineConfig({
         `,
       }
     ),
-    createSvg(path.join(path.resolve(__dirname, 'src/TUIRoom/assets/icons/svg/'), '/')),
-    AutoImport({
-      resolvers: [ElementPlusResolver()],
-    }),
-    Components({
-      resolvers: [ElementPlusResolver({
-        importStyle: 'sass',
-      })],
-    }),
   ],
   css: {
     preprocessorOptions: {